Initial import
This is a proof of concept for an image resizing microservice.  It
resizes images on demand and stores the cached images automatically.

There are still some safety mechanisms which should be put in place in
order to limit the chances of creating overly large images in the
backend.  We consider this beyond the scope of the current PoC.